Move component tests to component sub folder.
