+++ hyperspy/tests/component/test_component.py
