+++ hyperspy/tests/component/test_components.py
