+++ hyperspy/tests/component/test_components2D.py
